+++ index.html
@@ -3,7 +3,7 @@
 <head>
   <meta charset="utf-8">
   <script src="/js/functions.js"></script>
-  <script src="/js/main.js"></script>
+  <script src="/js/main.js" type="module"></script>
   <meta name="viewport" content="width=device-width,initial-scale=1">
   <link rel="stylesheet" href="css/normalize.css">
   <link rel="stylesheet" href="css/style.css">
